+++ dramatiq/middleware/__init__.py
@@ -1,5 +1,5 @@
 from .age_limit import AgeLimit  # noqa
-from .middleware import Middleware  # noqa
+from .middleware import Middleware, MiddlewareError, SkipMessage  # noqa
 from .prometheus import Prometheus  # noqa
 from .retries import Retries  # noqa
 from .time_limit import TimeLimit, TimeLimitExceeded  # noqa